misc: enlarge address space

+++ Utils/Common.h
@@ -23,7 +23,7 @@ enum {
 };
 
 const uint64_t FULLMASK = 0xFFFFFFFFFFFFFFFF;
-typedef uint16_t paddr_t;
+typedef uint32_t paddr_t;
 
 #define tlc_assert(cond, info) \
     do { \
